cleanup

+++ src/main/java/net/foxyas/changedaddon/procedures/LaethinPropertyValueProviderProcedure.java
@@ -3,7 +3,7 @@ package net.foxyas.changedaddon.procedures;
 import net.minecraft.world.item.ItemStack;
 
 public class LaethinPropertyValueProviderProcedure {
-    public static double execute(ItemStack itemstack) {
+    public static float execute(ItemStack itemstack) {
         if ((itemstack.getOrCreateTag().getString("type")).equals("dark_latex")) {
             return 1;
         }
